completes challenge 3
For a more challenging task, see if you can convert our data model from a class to a struct, then create an ObservableObject class wrapper around it that gets passed around. This will result in your class having one @Published property, which is the data struct inside it, and should make supporting Codable on the struct much easier.

Signed-off-by: Takasur Azeem <takasurazeem@gmail.com>